mv generate methods to component
